// src/app/page/master/satuan-standar/satuan-standar.interface.ts
import { EventEmitter } from '@angular/core';

export interface SatuanStandar {
	
	alamatEmail?: string;
	faksimile?: string;
	fixedPhone?: string;
	jamBuka?: string;
	jamTutup?: string;
	kdDepartemen?: string;
	kdKelasHead?: string;
	kdModulAplikasi?: string;
	kdPegawaiKepala?: string;
	kode?: any;
	kodeExternal?: string;
	lokasiRuangan?: string;
	mobilePhone?: string;
	namaExternal?: string;
	namaRuangan?: string;
	noCounter?: any;
	noRuangan?: string;
	prefixNoAntrian?: string;
	reportDisplay?: string;
	statusEnabled?: true;

	version?: any;
	noRec?: string;
	label?: any;
	value?: any;

}